example_code/news/requirements.txt
gunicorn
newsapi-python
Flask